Step7のレッスンファイルを追加

+++ Step6/Lesson1/Division.java
@@ -7,17 +7,17 @@ public class Division{
 
         //ユーザーに分子の入力を求める
         System.out.print("分子を入力してください: ");
-        int numerator = scanner.nextInt();
+        double numerator = scanner.nextDouble();
 
         //ユーザーに分母の入力を求める
         System.out.print("分母を入力してください: ");
-        int denominator = scanner.nextInt();
+        double denominator = scanner.nextDouble();
 
         //try-catchを使用した例外処理
         try{
             //割り算を行いを実行する
             //分母が0の場合はArithmeticExceptionをスローします
-            int result = numerator / denominator;
+            double result = numerator / denominator;
             System.out.println("結果: " + result);
         }catch (ArithmeticException e){
             //割る数（分母）が0の場合にエラー表示
